// src/app/types/black-lists.ts
export interface BlackListsTypes {
    Id: number;
    Adi: string;
    Soy: string;
    Aciklama: string;
    Tcno: string;
    Kimlik_no: string;
    Dogum_tarihi: string;
    Sistem_tarihi: string;
    Sistem_grubu: string;
    Otel_kodu: string | null;
    Ulke_xml: string;
    Kulanici: string | null;
    Acenta: string | null;
    "Xml Kodu": string;
    "ULke Adı": string;
}